+++ src/modules/lua_mtev_crypto.c
@@ -48,6 +48,9 @@
 #ifndef OPENSSL_STACK
 #define OPENSSL_STACK void
 #endif
+#ifndef X509V3_CTX_REPLACE
+#define X509V3_CTX_REPLACE 0x2
+#endif
 #endif
 
 #ifndef sk_OPENSSL_STRING_value
@@ -436,9 +439,7 @@ mtev_lua_crypto_rsa_gencsr(lua_State *L) {
     REQERR("crypto.rsa:gencsr could not set request version");
 
   lua_getfield(L,-1,"addext");
-#if OPENSSL_VERSION_NUMBER < 0x10100000L
-  if(!lua_isnil(L,-1)) REQERR("addext requires OpenSSL 1.1 or greater");
-#else
+
   X509V3_CTX ext_ctx;
   if(!lua_isnil(L,-1)) {
     if(!lua_istable(L,-1)) REQERR("addext value must be a table");
@@ -479,7 +480,6 @@ mtev_lua_crypto_rsa_gencsr(lua_State *L) {
       }
     }
   }
-#endif
   lua_pop(L,1);
 
   lua_getfield(L,-1,"subject");